bloom: use "true" division in the module
In Python 2.X, dividing two integers yields the integer floor of
the result by default: 1 / 2 = 0. Change the module to activate
"true" division, in which 1 / 2 = 0.5, by the use of a __future__
import. This only effects only place in the code where we had
to multiply some numbers by -1. to floatify them; there are other
places where we relied on the "classic" division behavior but
used the "//" operator which maintains that behavior even when
true division is activated. See PEP238 for more context.